refactor(#3038): extract QueuedPlaceholderState from SharedData (S1)
Lift cluster C — the queued-placeholder handoff state — out of the
discord/mod.rs giant into a new `shared_state` sibling module, following
the CoreState field-group precedent and the #3294/#3295 behaviour-
preserving decomposition standard.

What moves (verbatim):
- 3 fields (queued_placeholders, queue_exit_placeholder_clears,
  queued_placeholders_persist_locks) → `QueuedPlaceholderState`, replaced
  on `SharedData` by a single `queued: QueuedPlaceholderState` group field
  at the first member's original position.
- 8 inherent methods (queued_placeholders_persist_lock,
  insert/remove_queued_placeholder(_locked), queued_placeholder_still_owned,
  add/remove_pending_queue_exit_placeholder_clear(s)/_clear_one,
  pending_queue_exit_placeholder_clears) → a second `impl SharedData` block
  in shared_state.rs (inherent impls are valid from any module in the crate).

Body-edit equivalence (the only edits inside moved code):
1. `self.<field>` → `self.queued.<field>` — pure field-path re-wiring;
   method calls (`self.queued_placeholders_persist_lock(...)`) are NOT
   rewritten. Compile-time resolution only; control flow, await points,
   and lock acquire/release order unchanged.
2. `queued_placeholders_store::` → `super::queued_placeholders_store::` —
   same symbol, new module vantage.
3. 3 previously module-private methods widened to
   `pub(in crate::services::discord)` so the same mod.rs callers still
   resolve them across the new module boundary (reachability identical;
   compile-time-only re-annotation).

Constructor literals (3 sites: runtime_bootstrap.rs run_bot_build_shared_data,
mod.rs + voice_barge_in.rs test helpers) wrap the three consecutive member
inits in `queued: QueuedPlaceholderState { .. }` at the first member's
position. The init expressions — including run_bot_build_shared_data's
`load_queue_exit_placeholder_clears` disk read — stay byte-identical, so the
struct-literal text-order evaluation (load_generation / TurnFinalizer::spawn /
StatusPanelController::spawn / broadcast::channel relative order) is preserved.

The type is re-exported `pub(in crate::services::discord)` so the SharedData
surface stays frozen. Direct field accessors at 5 call sites
(gateway.rs, intake_gate.rs ×2, intake_turn.rs, runtime_bootstrap.rs helper)
get the mechanical `.queued.` insertion; accessor-method call sites are
untouched.

Invariants held:
- run_bot body (runtime_bootstrap.rs:895) byte-unchanged; only the
  independent helpers run_bot_build_shared_data and
  run_bot_spawn_recovery_and_flush_restart_reports are touched.
- #3016 hot files (turn_bridge/mod.rs, turn_finalizer.rs, recovery_engine.rs,
  tui_prompt_relay.rs) diff 0.
- round-5 P2 lock-span invariant intact (verbatim bodies; *_locked
  caller-holds-lock contract does not cross the module boundary).

Gates: fmt --check, cargo check --lib --tests (warnings as errors),
clippy --all-targets -D warnings, cargo test --lib services::discord
(1306 passed incl. the S0 characterization suite unchanged), audit
--check (mod.rs ratchet lowered 5188 → 4987; intake_turn/intake_gate/
runtime_bootstrap raised +1/+2/+2 for the mechanical re-wire, voice_barge_in
+0), check_await_holding_lock_ratchet (34), generate_inventory_docs +
check_agent_maintenance_docs synced.
