+++ ARCHITECTURE.md
@@ -530,6 +530,7 @@ src/
 │   │   ├── settings.rs
 │   │   ├── shadow_parity_warn.rs
 │   │   ├── shared_memory.rs
+│   │   ├── shared_state.rs
 │   │   ├── stall_recovery.rs
 │   │   ├── standby_relay.rs
 │   │   ├── status_panel_controller.rs
